Changed types in image
