+++ src/user/types/userTypes.ts
@@ -1,4 +1,5 @@
 import { ReactNode } from "react"
+import { FieldErrors, FieldValues, UseFormRegister } from "react-hook-form"
 
 interface BaseProps {
     id: string
@@ -27,4 +28,22 @@ export interface CardAvatarProps {
     children: ReactNode
     className?: string
     // style: string
+}
+
+export type AuthInputs = {
+    email: string
+    password: string
+    country: string
+    language: string
+    phone: string
+    firstName: string
+    lastName: string
+    image: File
+}
+
+export type UserFormProps = {
+    register: UseFormRegister<AuthInputs>,
+    errors: FieldErrors<AuthInputs>,
+    setValue: React.Dispatch<React.SetStateAction<string>>,
+    children: React.ReactNode
 }
